refactor(fuzz): split the validation generator into CLI and Maestro modules, mirrored in tests
